+++ docs/RELEASE-NOTES.md
@@ -1,15 +1,16 @@
 # Release Notes
 
-## 1.2.0 - ?
+## 2.0.0 - 21 January 2021
 
 Breaking changes:
-- (None)
+- Removed `EntityNotFoundException.EntityType` property.
 
 New features:
 - Added `DependencyNullException`.
 
 Bug fixes / internal changes:
-- (None)
+- Fixed `EntityNotFoundException` custom constructor to handle null params.
+- Fixed `EntityNotFoundException` to serialize `EntityId` property properly.
 
 ## 1.1.0 - 19 January 2021